+++ packages/widgetbook/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- **FIX**: Make addon-provided `InheritedWidget`s accessible via `BuildContext` in story `builder` and `setup` functions. ([#1894](https://github.com/widgetbook/widgetbook/pull/1894))
+
 ## 4.0.0-beta.3
 
 - **FIX**: Virtualize navigation tree to only build visible items, improving performance for large catalogs. ([#1886](https://github.com/widgetbook/widgetbook/pull/1886))